Adding Smart Home Code With Change configuration in MCAL and HAL layer

+++ MCAL/ADC/Adc.h
@@ -4,6 +4,6 @@
 #include <Std_Types.h>
 
 void Adc_Init();
-uint8 Adc_Read();
+uint16 Adc_Read();
 
 #endif /* MCAL_ADC_ADC_H_ */
